with generated <pre> tags
The former versions of this document were html from the wiki, run through pandoc.  The tables translated but the code blocks did not.

I've now written a perl script that scrubs out the wiki "code" tags from the html and replaces them with <pre></pre> tags, *then* this was run through stock html.  It doesn't preserve the titles of the code blocks, but otherwise I think this may work.